Prove algorithm. Refactoring.
