+++ src/main/scala/org/ergoplatform/modifiers/history/NiPoPowProof.scala
@@ -1,7 +1,6 @@
 package org.ergoplatform.modifiers.history
 
 import com.google.common.primitives.{Bytes, Ints}
-import org.ergoplatform.mining.difficulty.RequiredDifficulty
 import org.ergoplatform.modifiers.ErgoPersistentModifier
 import org.ergoplatform.settings.Algos
 import scorex.core.ModifierTypeId
@@ -18,11 +17,11 @@ case class NiPoPowProof(m: Int,
                         sizeOpt: Option[Int] = None)
   extends ErgoPersistentModifier with ModifierValidator {
 
-  import NiPoPowProof._
+  import NiPoPowAlgos._
 
   override type M = NiPoPowProof
 
-  override val modifierTypeId: ModifierTypeId = TypeId
+  override val modifierTypeId: ModifierTypeId = NiPoPowProof.TypeId
 
   override def serializedId: Array[Byte] = Algos.hash(bytes)
 
@@ -51,114 +50,7 @@ case class NiPoPowProof(m: Int,
 }
 
 object NiPoPowProof {
-
   val TypeId: ModifierTypeId = ModifierTypeId @@ (110: Byte)
-
-  def maxLevelOf(header: Header): Int = {
-    if (!header.isGenesis) {
-      def log2(x: Double) = math.log(x) / math.log(2)
-      val requiredTarget = org.ergoplatform.mining.q / RequiredDifficulty.decodeCompactBits(header.nBits)
-      val realTarget = header.powSolution.d
-      val level = log2(requiredTarget.doubleValue) - log2(realTarget.doubleValue)
-      level.toInt
-    } else {
-      Int.MaxValue
-    }
-  }
-
-  def bestArg(chain: Seq[Header])(m: Int): Int = {
-    def loop(level: Int, acc: Seq[(Int, Int)] = Seq.empty): Seq[(Int, Int)] = {
-      if (level == 0) {
-        loop(level + 1, acc :+ (0, chain.size)) // Supposing each header is at least of level 0.
-      } else {
-        val args = chain.filter(maxLevelOf(_) >= level)
-        if (args.lengthCompare(m) >= 0) loop(level + 1, acc :+ (level, args.size)) else acc
-      }
-    }
-    loop(level = 0).map { case (lvl, size) =>
-      math.pow(2, lvl) * size // 2^µ * |C↑µ|
-    }.max.toInt
-  }
-
-  def lowestCommonAncestor(leftChain: Seq[Header], rightChain: Seq[Header]): Option[Header] = {
-    def lcaIndex(startIdx: Int): Int = {
-      if (leftChain.lengthCompare(startIdx) >= 0 && rightChain.lengthCompare(startIdx) >= 0 &&
-        leftChain(startIdx) == rightChain(startIdx)) {
-        lcaIndex(startIdx + 1)
-      } else {
-        startIdx - 1
-      }
-    }
-    if (leftChain.headOption.exists(h => rightChain.headOption.contains(h))) Some(leftChain(lcaIndex(1))) else None
-  }
-
-  def updateInterlinks(header: Header): Seq[ModifierId] = {
-    if (!header.isGenesis) {
-      val genesis = header.interlinks.head
-      val tail = header.interlinks.tail
-      val prevLevel = maxLevelOf(header)
-      if (prevLevel > 0) {
-        (genesis +: tail.dropRight(prevLevel)) ++ Seq.fill(prevLevel)(header.id)
-      } else {
-        header.interlinks
-      }
-    } else {
-      Seq(header.id)
-    }
-  }
-
-  def goodSuperChain(chain: Seq[Header], superChain: Seq[Header], level: Int): Boolean = {
-    import org.ergoplatform.settings.Constants.NiPoPowParams._
-    superChainQuality(chain, superChain, level)(m, d) && multiLevelQuality(chain, superChain, level)(k1, d)
-  }
-
-  private def locallyGood(chainSize: Int, superChainSize: Int, level: Int)(d: Float): Boolean = {
-    superChainSize > (1 - d) * math.pow(2, -level) * chainSize
-  }
-
-  /** @param chain      - Full chain (C)
-    * @param superChain - Super-chain of level µ (C↑µ)
-    * @param level      - Level of super-chain (µ) */
-  private def superChainQuality(chain: Seq[Header], superChain: Seq[Header], level: Int)(m: Int, d: Float): Boolean = {
-    val downChain = chain.dropWhile(_ == superChain.head).takeWhile(_ == superChain.last) // C[C↑µ[0]:C↑µ[−1]], or C'↓
-    def checkLocalGoodnessAt(mValue: Int): Boolean = {
-      mValue match {
-        case mToTest if mToTest < chain.size &&
-          locallyGood(math.min(superChain.size, mToTest), math.min(downChain.size, mToTest), level)(d) =>
-          checkLocalGoodnessAt(mToTest + 1)
-        case mToTest if mToTest < chain.size =>
-          false
-        case _ =>
-          true
-      }
-    }
-    checkLocalGoodnessAt(m)
-  }
-
-  private def multiLevelQuality(chain: Seq[Header], superChain: Seq[Header], level: Int)(k1: Int, d: Float): Boolean = {
-    val downChain = chain.dropWhile(_ == superChain.head).takeWhile(_ == superChain.last) // C'↓
-    def checkQualityAt(levelToCheck: Int): Boolean = {
-      levelToCheck match {
-        case lvl if lvl > 0 =>
-          val subChain = downChain.filter(maxLevelOf(_) >= lvl - 1) // C* = C'↓↑µ'−1
-          if (subChain.nonEmpty) {
-            val upperSubChainSize = subChain.count(maxLevelOf(_) >= lvl) // |C*↑µ'|
-            if (upperSubChainSize >= k1 &&
-              !(subChain.count(maxLevelOf(_) >= level) >= (1 - d) * math.pow(2, level - lvl) * upperSubChainSize)) {
-              false
-            } else {
-              checkQualityAt(lvl - 1)
-            }
-          } else {
-            checkQualityAt(lvl - 1)
-          }
-        case _ =>
-          true
-      }
-    }
-    checkQualityAt(level)
-  }
-
 }
 
 object NiPoPowProofSerializer extends Serializer[NiPoPowProof] {
