+++ src/main/scala/org/ergoplatform/settings/Constants.scala
@@ -58,12 +58,4 @@ object Constants {
       BlockTransactions.modifierTypeId -> BlockTransactionsSerializer,
       ADProofs.modifierTypeId -> ADProofSerializer,
       Transaction.ModifierTypeId -> ErgoTransactionSerializer)
-
-  object NiPoPowParams {
-    val m = 30
-    val k = 30
-    val d = 0.45f
-    val k1: Int = k
-  }
-
 }
