+++ src/main/scala/org/ergoplatform/settings/NiPoPowSettings.scala
@@ -1,0 +1,3 @@
+package org.ergoplatform.settings
+
+case class NiPoPowSettings(m: Int, k: Int, k1: Int, d: Float)
