+++ src/main/scala/org/ergoplatform/settings/NodeConfigurationSettings.scala
@@ -1,6 +1,7 @@
 package org.ergoplatform.settings
 
 import net.ceedubs.ficus.Ficus._
+import net.ceedubs.ficus.readers.ArbitraryTypeReader._
 import net.ceedubs.ficus.readers.ValueReader
 import org.ergoplatform.nodeView.state.StateType
 
@@ -19,7 +20,8 @@ case class NodeConfigurationSettings(stateType: StateType,
                                      mining: Boolean,
                                      miningDelay: FiniteDuration,
                                      offlineGeneration: Boolean,
-                                     keepVersions: Int)
+                                     keepVersions: Int,
+                                     poPowSettings: NiPoPowSettings)
 
 
 trait NodeConfigurationReaders extends StateTypeReaders with ModifierIdReader {
@@ -27,14 +29,18 @@ trait NodeConfigurationReaders extends StateTypeReaders with ModifierIdReader {
   implicit val nodeConfigurationReader: ValueReader[NodeConfigurationSettings] = { (cfg, path) =>
     val stateTypeKey = s"$path.stateType"
     val stateType = stateTypeFromString(cfg.as[String](stateTypeKey), stateTypeKey)
-    NodeConfigurationSettings(stateType,
-                              cfg.as[Boolean](s"$path.verifyTransactions"),
-                              cfg.as[Int](s"$path.blocksToKeep"),
-                              cfg.as[Boolean](s"$path.PoPoWBootstrap"),
-                              cfg.as[Int](s"$path.minimalSuffix"),
-                              cfg.as[Boolean](s"$path.mining"),
-                              cfg.as[FiniteDuration](s"$path.miningDelay"),
-                              cfg.as[Boolean](s"$path.offlineGeneration"),
-                              cfg.as[Int](s"$path.keepVersions"))
+    NodeConfigurationSettings(
+      stateType,
+      cfg.as[Boolean](s"$path.verifyTransactions"),
+      cfg.as[Int](s"$path.blocksToKeep"),
+      cfg.as[Boolean](s"$path.PoPoWBootstrap"),
+      cfg.as[Int](s"$path.minimalSuffix"),
+      cfg.as[Boolean](s"$path.mining"),
+      cfg.as[FiniteDuration](s"$path.miningDelay"),
+      cfg.as[Boolean](s"$path.offlineGeneration"),
+      cfg.as[Int](s"$path.keepVersions"),
+      cfg.as[NiPoPowSettings](s"$path.niPoPow")
+    )
   }
+
 }
